Move public to packages/legacy.
