+++ packages/legacy/public/icons.woff2
