+++ packages/legacy/tailwind.config.ts
@@ -1,7 +1,7 @@
 import type { Config } from 'tailwindcss';
-import { theme } from '../../public/theme.js';
+import { theme } from './public/theme.js';
 
-export { theme } from '../../public/theme.js';
+export { theme } from './public/theme.js';
 
 export default {
   content: ['./index.html', './src/**/*.{ts,vue,svelte,component}'],
