+++ packages/legacy/vite.config.ts
@@ -54,5 +54,4 @@ export default defineConfig({
     port: 5174,
     host: true,
   },
-  publicDir: path.resolve(__dirname, '../../public'),
 });
